fixed instance offset

+++ nmea-router/src/main/java/org/vesalainen/nmea/router/endpoint/n2kgw/SourceManager.java
@@ -94,7 +94,7 @@ public class SourceManager extends JavaLogging
     public int getInstanceOffset(int canId)
     {
         int sa = PGN.sourceAddress(canId);
-        Source src = map.get(sa);
+        Source src = map.get((byte)sa);
         if (src != null)
         {
             return src.instanceOffset;
